+++ CHANGELOG.md
@@ -71,6 +71,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingTokenDesign resource
 - IssuingTokenRequest sub-resource
 - IssuingTokenActivation sub-resource
+- IssuingStockRule resource
 ### Changed
 - IssuingBin resource to IssuingProduct
 - settlement parameter to funding_type and client parameter to holder_type of IssuingProduct